Add restrictors
In a group without access to messages, bot cannot see when a user types
in a module code after a successful /add_group. To resolve this, we
limit commands /add_group and /remove_group (because I don't wanna deal
with figuring out who clicked the InlineKeyboard button in a group.

The restrictors come in the form of decorators. Good practice is to
define 'lower level' or 'less abstracted' functions at the bottom of
the file. This means that the @decorator syntax would come before the
actual definition of the decorator, which obviously doesn't work---the
decorator cannot decorate a function, as it is not yet defined.

To resolve this, we abstract all 'helper functions', including the
restrictors into a separate module, utils. There are two minor problems
with this, however:
    (1) Some functions require access to the database. They used to
        access the db variable in the global scope, but now must take
        the db as an positional argument. Not going to include db in
        utils itself seems to be outside the scope of 'utilities'.
    (2) There are MANY functions in utils, all of which are in use in
        the module commands. This means that we use
            from utils import *
        which obviously isn't ideal (implicit namespaces).